In LeagueOfLegendsCreationUtils: edited the createGameStats method to only take arguments Game game and float Duration. In LeagueOfLegendsResourceImpl: a GameStats object is created right after game is created if that game does not have a GameStats object linked to it already. GameStats is first created with only a reference to a specific game and the duration of said game. Outside the loop where csv file is read, after gameTeamStats is created and populated, the rest of gameStats is also populated

+++ tdt4250.project.LeagueOfLegends/src/leagueOfLegends/GameStats.java
@@ -64,12 +64,12 @@ public interface GameStats extends EObject {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Duration</em>' attribute.
-	 * @see #setDuration(int)
+	 * @see #setDuration(float)
 	 * @see leagueOfLegends.LeagueOfLegendsPackage#getGameStats_Duration()
 	 * @model
 	 * @generated
 	 */
-	int getDuration();
+	float getDuration();
 
 	/**
 	 * Sets the value of the '{@link leagueOfLegends.GameStats#getDuration <em>Duration</em>}' attribute.
@@ -79,7 +79,7 @@ public interface GameStats extends EObject {
 	 * @see #getDuration()
 	 * @generated
 	 */
-	void setDuration(int value);
+	void setDuration(float value);
 
 	/**
 	 * Returns the value of the '<em><b>Winner</b></em>' attribute.
